+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "cSpell.words": [
+        "Hodl",
+        "MINICHEFV",
+        "SUSHISWAP",
+        "WMATIC"
+    ]
+}
